+++ packages/react-ui/__tests__/widgets/CategoryWidgetUI.test.js
@@ -168,4 +168,27 @@ describe('CategoryWidgetUI', () => {
       expect(screen.getByText('Others (4)')).toBeInTheDocument();
     });
   });
+
+  describe('when data name is not string', () => {
+    const NotStringData = [{
+      name: true,
+      value: 100
+    }, {
+      name: false,
+      value: 101
+    }]
+    test('should render properly', () => {
+      render(<CategoryWidgetUI data={NotStringData} />);
+  
+      expect(screen.getByText(/true/)).toBeInTheDocument();
+      expect(screen.getByText(/101/)).toBeInTheDocument();
+    });
+
+    test('should maintain typing when filters', () => {
+      const onSelectedCategoriesChangeFn = jest.fn()
+      render(<CategoryWidgetUI data={NotStringData} onSelectedCategoriesChange={onSelectedCategoriesChangeFn} />);
+      userEvent.click(screen.getByText(/true/))
+      expect(onSelectedCategoriesChangeFn).toHaveBeenCalledWith([true])
+    })
+  })
 });