+++ packages/react-ui/src/types.d.ts
@@ -11,7 +11,7 @@ export type WrapperWidgetUI = {
   children?: React.ReactNode;
 };
 
-export type CategoryWidgetUIData = { name: string; value: number }[];
+export type CategoryWidgetUIData = { name: number | string | boolean; value: number }[];
 export type CategoryWidgetUI = {
   data: CategoryWidgetUIData;
   isLoading?: boolean;
@@ -44,6 +44,24 @@ export type HistogramWidgetUI = {
   name?: string;
   onSelectedBarsChange?: Function;
   height?: number;
+};
+
+export type BarWidgetUI = {
+  xAxisData: (string | number)[];
+  yAxisData: (string | number)[] | (string | number)[][];
+  series?: string[]
+  colors?: string | string[];
+  stacked?: boolean;
+  labels?: object;
+  tooltip?: boolean;
+  tooltipFormatter?: Function;
+  xAxisFormatter?: Function;
+  yAxisFormatter?: Function;
+  selectedBars?: number[] | [number, number][];
+  onSelectedBarsChange?: Function;
+  height?: string | number;
+  filterable?: boolean;
+  animation?: boolean;
 };
 
 export type PieWidgetUIData = { name: string; value: number }[];